+++ logprep/processor/list_comparison/processor.py
@@ -26,15 +26,19 @@ Processor Configuration
 .. automodule:: logprep.processor.list_comparison.rule
 """
 
+import logging
+
 from attrs import define, field, validators
 
-from logprep.abc.processor import Processor
+from logprep.abc.processor import Processor, ProcessorResult
 from logprep.processor.list_comparison.rule import ListComparisonRule
 from logprep.util.helper import (
     add_fields_to,
     get_dotted_field_value,
     join_dotted_fields,
 )
+
+logger = logging.getLogger("ListComparison")
 
 
 class ListComparison(Processor):
@@ -56,7 +60,11 @@ class ListComparison(Processor):
     def setup(self):
         super().setup()
         for rule in self.rules:
-            rule.init_list_comparison(self._config.list_search_base_path)
+            try:
+                rule.init_list_comparison(self._config.list_search_base_path)
+            except Exception as ex:
+                self._failed_init_rules[rule] = ex
+                logger.warning("Failed to initialize list comparison rule: %s", ex)
 
     def _apply_rules(self, event, rule):
         """